requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
pillow
pandas
numpy
plotly
